+++ java-client/src/it/scala/OrderIntegrationSpec.scala
@@ -17,15 +17,23 @@ class OrderIntegrationSpec extends WordSpec with MustMatchers {
     "change the payment state of an order" in {
       val order = newOrderOf1Product
       order.getPaymentState must be(null)
-      val updatedOrder = client.orders.updatePaymentState(order.getIdAndVersion, CreditOwed).execute()
+      val updatedOrder = client.orders.updateOrder(order.getIdAndVersion, new OrderUpdate().setPaymentState(CreditOwed)).execute()
       updatedOrder.getPaymentState must be(CreditOwed)
     }
 
     "change the shipment state of an order" in {
       val order = newOrderOf1Product
       order.getShipmentState must be (null)
-      val updatedOrder = client.orders.updateShipmentState(order.getIdAndVersion, Shipped).execute()
+      val updatedOrder = client.orders.updateOrder(order.getIdAndVersion, new OrderUpdate().setShipmentState(Shipped)).execute()
       updatedOrder.getShipmentState must be (Shipped)
+    }
+
+    "add tracking data" in {
+      val order = newOrderWithShippingMethod
+      val data = new TrackingData("xlf56b", "SPHERE parcels")
+      val updatedOrder = client.orders.updateOrder(order.getIdAndVersion, new OrderUpdate().addTrackingData(data)).execute()
+      val actual = updatedOrder.getShippingInfo.getTrackingData.get(0)
+      actual must be(data)
     }
   }
 }